updated Vela

+++ windows/SRSDG645.py
@@ -6,9 +6,10 @@ Created on Thu Jan 17 13:51:25 2019
 @author: robert
 """
 
-from PyQt4 import QtCore, QtGui, uic
-from PyQt4.QtCore import (pyqtSlot, QThread, pyqtSignal)
-from PyQt4.QtGui import (QPixmap, QImage, QLabel)
+from PyQt5 import QtCore, QtGui, uic
+from PyQt5.QtCore import (pyqtSlot, QThread, pyqtSignal)
+from PyQt5.QtGui import (QPixmap, QImage)
+from PyQt5.QtWidgets import QLabel
 import numpy as np
 import threading
 import os
@@ -124,15 +125,15 @@ class DGWindow(QtBaseClass, Ui_DGWindow):
             Identifier for the channel, either T0, T1, or A-H.
         """
         delay = self.settings[channel]['delay']
-        s = np.floor(delay)
+        s = int(np.floor(delay))
         delay = (delay - s)*1000
-        ms = np.floor(delay)
+        ms = int(np.floor(delay))
         delay = (delay - ms)*1000
-        us = np.floor(delay)
+        us = int(np.floor(delay))
         delay = (delay - us)*1000
-        ns = np.floor(delay)
+        ns = int(np.floor(delay))
         delay = (delay - ns)*1000
-        ps = np.floor(delay)
+        ps = int(np.floor(delay))
         self.sField.setValue(s)
         self.msField.setValue(ms)
         self.usField.setValue(us)
